restructure `tracing-glog` to prioritize `chrono`
