+++ src/format.rs
@@ -1,52 +1,10 @@
 #[cfg(feature = "ansi")]
-use crate::nu_ansi_term::{Color, Style};
-use std::{fmt, io};
-use time::{format_description::FormatItem, formatting::Formattable, OffsetDateTime};
+use nu_ansi_term::{Color, Style};
+use std::fmt;
 use tracing::{Level, Metadata};
 use tracing_subscriber::fmt::{format::Writer, time::FormatTime};
 
-#[cfg(feature = "chrono")]
 use tracing_subscriber::fmt::time::{ChronoLocal, ChronoUtc};
-
-/// A bridge between `fmt::Write` and `io::Write`.
-///
-/// This is used by the timestamp formatting implementation for the `time`
-/// crate and by the JSON formatter. In both cases, this is needed because
-/// `tracing-subscriber`'s `FormatEvent`/`FormatTime` traits expect a
-/// `fmt::Write` implementation, while `serde_json::Serializer` and `time`'s
-/// `format_into` methods expect an `io::Write`.
-pub(crate) struct WriteAdaptor<'a> {
-    fmt_write: &'a mut dyn fmt::Write,
-}
-
-impl<'a> WriteAdaptor<'a> {
-    pub(crate) fn new(fmt_write: &'a mut dyn fmt::Write) -> Self {
-        Self { fmt_write }
-    }
-}
-
-impl<'a> io::Write for WriteAdaptor<'a> {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        let s =
-            std::str::from_utf8(buf).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-        self.fmt_write
-            .write_str(s)
-            .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
-
-        Ok(s.as_bytes().len())
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        Ok(())
-    }
-}
-
-impl<'a> fmt::Debug for WriteAdaptor<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.pad("WriteAdaptor { .. }")
-    }
-}
 
 pub(crate) struct FmtLevel {
     pub level: Level,
@@ -94,49 +52,6 @@ impl fmt::Display for FmtLevel {
     }
 }
 
-/// Formats the current [UTC time] using a [formatter] from the [`time` crate].
-///
-/// To format the current [local time] instead, use the [`LocalTime`] type.
-///
-/// [UTC time]: time::OffsetDateTime::now_utc
-/// [formatter]: time::formatting::Formattable
-/// [`time` crate]: time
-/// [local time]: time::OffsetDateTime::now_local
-#[derive(Clone, Debug)]
-pub struct UtcTime<F = Vec<FormatItem<'static>>> {
-    format: F,
-}
-
-impl<F> FormatTime for UtcTime<F>
-where
-    F: Formattable,
-{
-    fn format_time(&self, writer: &mut Writer<'_>) -> fmt::Result {
-        let now = OffsetDateTime::now_utc();
-
-        #[cfg(feature = "ansi")]
-        if writer.has_ansi_escapes() {
-            let style = Style::new().dimmed();
-            write!(writer, "{}", style.prefix())?;
-            format_datetime(writer, now, &self.format)?;
-            write!(writer, "{}", style.suffix())?;
-            return Ok(());
-        }
-
-        format_datetime(writer, now, &self.format)
-    }
-}
-
-impl Default for UtcTime {
-    fn default() -> Self {
-        let format: Vec<FormatItem> = time::format_description::parse(
-            "[month][day] [hour]:[minute]:[second].[subsecond digits:6]",
-        )
-        .expect("Unable to make time formatter");
-        Self { format }
-    }
-}
-
 /// Formats the current [UTC time] using [`chrono` crate].
 ///
 /// To format the current local time instead, use the [`ChronoLocalTime`]
@@ -144,13 +59,11 @@ impl Default for UtcTime {
 ///
 /// [UTC time]: ChronoUtc
 /// [`chrono` crate]: chrono
-#[cfg(feature = "chrono")]
 #[derive(Clone, Debug)]
 pub struct ChronoUtcTime {
     time: ChronoUtc,
 }
 
-#[cfg(feature = "chrono")]
 impl FormatTime for ChronoUtcTime {
     fn format_time(&self, w: &mut Writer<'_>) -> fmt::Result {
         #[cfg(feature = "ansi")]
@@ -166,68 +79,12 @@ impl FormatTime for ChronoUtcTime {
     }
 }
 
-#[cfg(feature = "chrono")]
 impl Default for ChronoUtcTime {
     fn default() -> Self {
         let fmt_string = String::from("%m%d %H:%M:%S%.6f");
         Self {
             time: ChronoUtc::new(fmt_string),
         }
-    }
-}
-
-/// Formats the current [local time] using a [formatter] from the [`time` crate].
-///
-/// To format the current [UTC time] instead, use the [`UtcTime`] type.
-///
-/// <div class="example-wrap" style="display:inline-block">
-/// <pre class="compile_fail" style="white-space:normal;font:inherit;">
-///     <strong>Warning</strong>: The <a href = "https://docs.rs/time/0.3/time/"><code>time</code>
-///     crate</a> must be compiled with <code>--cfg unsound_local_offset</code> in order to use
-///     local timestamps. When this cfg is not enabled, local timestamps cannot be recorded, and
-///     events will be logged without timestamps.
-///
-///    See the <a href="https://docs.rs/time/0.3.4/time/#feature-flags"><code>time</code>
-///    documentation</a> for more details.
-/// </pre></div>
-///
-/// [local time]: time::OffsetDateTime::now_local
-/// [formatter]: time::formatting::Formattable
-/// [`time` crate]: time
-/// [UTC time]: time::OffsetDateTime::now_utc
-#[derive(Clone, Debug)]
-pub struct LocalTime<F = Vec<FormatItem<'static>>> {
-    format: F,
-}
-
-impl Default for LocalTime {
-    fn default() -> Self {
-        let format: Vec<FormatItem> = time::format_description::parse(
-            "[month][day] [hour]:[minute]:[second].[subsecond digits:6]",
-        )
-        .expect("Unable to make time formatter");
-        Self { format }
-    }
-}
-
-impl<F> FormatTime for LocalTime<F>
-where
-    F: Formattable,
-{
-    fn format_time(&self, writer: &mut Writer<'_>) -> fmt::Result {
-        let now = OffsetDateTime::now_local().map_err(|_| fmt::Error)?;
-
-        #[cfg(feature = "ansi")]
-        if writer.has_ansi_escapes() {
-            let style = Style::new().dimmed();
-            write!(writer, "{}", style.prefix())?;
-            format_datetime(writer, now, &self.format)?;
-            // necessary to provide space between the time and the PID
-            write!(writer, "{}", style.suffix())?;
-            return Ok(());
-        }
-
-        format_datetime(writer, now, &self.format)
     }
 }
 
@@ -238,12 +95,10 @@ where
 ///
 /// [`local time`]: ChronoLocal
 /// [`chrono` crate]: chrono
-#[cfg(feature = "chrono")]
 pub struct ChronoLocalTime {
     time: ChronoLocal,
 }
 
-#[cfg(feature = "chrono")]
 impl FormatTime for ChronoLocalTime {
     fn format_time(&self, w: &mut Writer<'_>) -> fmt::Result {
         #[cfg(feature = "ansi")]
@@ -259,7 +114,6 @@ impl FormatTime for ChronoLocalTime {
     }
 }
 
-#[cfg(feature = "chrono")]
 impl Default for ChronoLocalTime {
     fn default() -> Self {
         let fmt_string = String::from("%m%d %H:%M:%S%.6f");
@@ -267,17 +121,6 @@ impl Default for ChronoLocalTime {
             time: ChronoLocal::new(fmt_string),
         }
     }
-}
-
-fn format_datetime(
-    into: &mut Writer<'_>,
-    now: OffsetDateTime,
-    fmt: &impl Formattable,
-) -> fmt::Result {
-    let mut into = WriteAdaptor::new(into);
-    now.format_into(&mut into, fmt)
-        .map_err(|_| fmt::Error)
-        .map(|_| ())
 }
 
 pub(crate) struct FormatProcessData<'a> {
