+++ src/lib.rs
@@ -93,10 +93,14 @@
 #[deny(rustdoc::broken_intra_doc_links)]
 mod format;
 
+#[cfg(feature = "time")]
+pub mod time_crate;
+
 #[cfg(feature = "ansi")]
 mod nu_ansi_term {
     pub use ::nu_ansi_term::*;
 }
+
 #[cfg(not(feature = "ansi"))]
 mod nu_ansi_term {
     // Minimal API shim for nu_ansi_term to avoid a pile of #[cfg(feature = "ansi")] directives.
@@ -124,9 +128,7 @@ mod nu_ansi_term {
 
 use crate::nu_ansi_term::Style;
 use format::FmtLevel;
-#[cfg(feature = "chrono")]
 pub use format::{ChronoLocalTime, ChronoUtcTime};
-pub use format::{LocalTime, UtcTime};
 use std::fmt;
 use tracing::{
     field::{Field, Visit},
@@ -147,7 +149,7 @@ use crate::format::{FormatProcessData, FormatSpanFields};
 /// A [glog]-inspired span and event formatter.
 ///
 /// [glog]: https://github.com/google/glog
-pub struct Glog<T = UtcTime> {
+pub struct Glog<T = ChronoUtcTime> {
     timer: T,
     with_span_context: bool,
     with_thread_names: bool,
@@ -239,10 +241,10 @@ impl<T> Glog<T> {
     }
 }
 
-impl Default for Glog<UtcTime> {
+impl Default for Glog<ChronoUtcTime> {
     fn default() -> Self {
         Glog {
-            timer: UtcTime::default(),
+            timer: ChronoUtcTime::default(),
             with_thread_names: false,
             with_target: false,
             with_span_context: true,
